set paper to f4

+++ resources/views/admin/santri/components/actions.blade.php
@@ -1,4 +1,6 @@
 <div class="btn-group">
     <button class="btn btn-sm btn-primary" onclick="editCustomer({{ $Santri->id }})">Edit</button>
-    <button class="btn btn-sm btn-danger " onclick="deleteCustomers({{ $Santri->id }})">Delete</button>
+    @if (Auth::user()->role == 'Admin')
+        <button class="btn btn-sm btn-danger " onclick="deleteCustomers({{ $Santri->id }})">Delete</button>
+    @endif
 </div>
